Make shared dandisetListing vuex module
