+++ web/src/store/index.js
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 
 import girder from '@/store/girder';
-import myDandisets from '@/store/myDandisets';
+import dandisetListing from '@/store/dandisetListing';
 import stats from '@/store/stats';
 
 Vue.use(Vuex);
@@ -11,8 +11,8 @@ Vue.use(Vuex);
 export default new Vuex.Store({
   modules: {
     girder,
-    publicDandisets: myDandisets,
-    myDandisets,
+    publicDandisets: dandisetListing(),
+    myDandisets: dandisetListing({ user: true }),
     stats,
   },
 });
